Add unigram and vocabulary with the new table structure.
Add Ngram class which encapsulates manipulation over Ngrams.
Add `insert_into_ngram` generic function which is meant to be used for any insertions into ngrams.
Add an after generation cleanup, which fills all missing cells with 0.0 and appends `Sum` column.
The code in this PR is just to showcase the approach. Multiple parts are not yet properly optimized.