Fix bug with prior prob
The code calculating the prior probability assumed that the language it was calculating the probability for was present in the test data. this lead to a bug where test_data[DF_COLUMN_LANG].value_counts()[lang] would crash the program, since the key lang didn't exist. The code was adjusted to account for this case.